Framework: Add WatchIgnorePlugin for node_modules
This change cuts down the numbers of watchers to set from over 7000 to 500-600.

We are setting watchers for all files in node_modules by default but since these files don't change often
and if they change they usually warrant a full restart of the dev server, they are OK to ignore. And by ignoring that folder,
we no longer exhaust the available file handles in the OS.